restructuring the code, deleting some useless components, adding a favicon
