--- src/main/webapp/app/_bootstrap-variables.scss
@@ -1,6 +1,0 @@
-/*
-* Bootstrap overrides https://v4-alpha.getbootstrap.com/getting-started/options/
-* All values defined in bootstrap source
-* https://github.com/twbs/bootstrap/blob/v4-dev/scss/_variables.scss can be overwritten here
-* Make sure not to add !default to values here
-*/
